XRDDEV-746 Move more views
