+++ src/proxy-ui-api/frontend/src/router.ts
@@ -13,21 +13,21 @@ import Settings from '@/views/Settings/Settings.vue';
 import SystemParameters from '@/views/Settings/SystemParameters.vue';
 import BackupAndRestore from '@/views/Settings/BackupAndRestore.vue';
 import Diagnostics from '@/views/Diagnostics/Diagnostics.vue';
-import AddSubsystem from '@/views/AddSubsystem.vue';
-import AddClient from '@/views/AddClient.vue';
+import AddSubsystem from '@/views/AddSubsystem/AddSubsystem.vue';
+import AddClient from '@/views/AddClient/AddClient.vue';
 import Subsystem from '@/views/Clients/Subsystem.vue';
 import ClientDetails from '@/views/Clients/Details/ClientDetails.vue';
 import InternalServers from '@/views/Clients/InternalServers/InternalServers.vue';
 import Services from '@/views/Clients/Services/Services.vue';
 import ServiceClients from '@/views/Clients/ServiceClients/ServiceClients.vue';
 import LocalGroups from '@/views/Clients/LocalGroups/LocalGroups.vue';
-import ClientTlsCertificate from '@/views/ClientTlsCertificate.vue';
+import ClientTlsCertificate from '@/views/ClientTlsCertificate/ClientTlsCertificate.vue';
 import AppError from '@/views/AppError.vue';
 import LocalGroup from '@/views/LocalGroup/LocalGroup.vue';
-import ServiceDescriptionDetails from '@/views/ServiceDescriptionDetails.vue';
-import TokenDetails from '@/views/TokenDetails.vue';
-import KeyDetails from '@/views/KeyDetails.vue';
-import CertificateDetails from '@/views/CertificateDetails.vue';
+import ServiceDescriptionDetails from '@/views/ServiceDescriptionDetails/ServiceDescriptionDetails.vue';
+import TokenDetails from '@/views/TokenDetails/TokenDetails.vue';
+import KeyDetails from '@/views/KeyDetails/KeyDetails.vue';
+import CertificateDetails from '@/views/CertificateDetails/CertificateDetails.vue';
 import Service from '@/views/Service/Service.vue';
 import store from '@/store';
 import { RouteName, Permissions } from '@/global';
