+++ src/proxy-ui-api/frontend/src/views/CertificateDetails/CertificateDetails.vue
