+++ src/proxy-ui-api/frontend/src/views/ClientTlsCertificate/ClientTlsCertificate.vue
