+++ src/proxy-ui-api/frontend/src/views/TokenDetails/TokenDetails.vue
@@ -114,6 +114,6 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/detail-views';
+@import '../../assets/detail-views';
 </style>
